# Java: rewrite receive callback