… orte/libjorte/Makefile.omk
… orte/libjorte/setRecvInfo.c