@wentasah 2015-06-10
Could you please fix code style errors reported at
http://compile.felk.cvut.cz:8011/builders/uncrustify/builds/70/steps/diff/logs/stdio?

I've setup travis-ci so that it will automatically checks the pull requests (PR). If it does not check after push to this PR (because if was created before Travis was set up), then create a new PR.

Thanks.