@Vajnar 2015-06-10
Pushed coding style fixes.